update graphene, graphql dependancies
