+++ requirements.txt
@@ -42,9 +42,9 @@ future==0.16.0
 futures==3.1.1
 google-i18n-address==2.0.1
 google-measurement-protocol==0.1.6
-graphene==1.1.2
-graphene-django==1.2.0
-graphql-core==1.0.1
+graphene==2.0.1
+graphene-django==2.0.0
+graphql-core==2.0
 graphql-relay==0.4.5
 html5lib==1.0b8
 idna==2.2
@@ -63,7 +63,7 @@ pbr==1.10.0
 Pillow==3.4.2
 premailer==3.0.1
 prices==0.5.8
-promise==1.0.0
+promise==2.1 # via graphene, graphene-django, graphql-core, graphql-relay
 psycopg2==2.6.2
 purl==1.3
 py==1.4.32
