+++ saleor/return_purchase/templates/return_purchase/form.html
@@ -12,7 +12,7 @@
 {% endblock %}
 
  {% block breadcrumbs %}
-     <li><a href="/return/sale/">Returns Management</a></li>
+     <li><a href="/return/purchase/">Returns Management</a></li>
      <li class=""><a href="/return/purchase/" >Return Purchase</a></li>
      <li class="active">Return Purchase</li>
  {% endblock %}
