+++ saleor/return_purchase/templates/return_purchase/items.html
@@ -8,7 +8,7 @@
 {% endblock %}
 
  {% block breadcrumbs %}
-     <li ><a href="/return/sale/">Returns Management</a></li>
+     <li ><a href="/return/purchase/">Returns Management</a></li>
      <li class=""><a href="/return/purchase/"> Return Purchase</a></li>
      <li class="active"><a >Returned Items</a></li>
  {% endblock %}
