+++ saleor/return_purchase/templates/return_purchase/list.html
@@ -8,7 +8,7 @@
 {% endblock %}
 
  {% block breadcrumbs %}
-     <li><a href="/return/sale/">Returns Management</a></li>
+     <li><a href="/return/purchase/">Returns Management</a></li>
      <li class="active"><a> Return Purchases</a></li>
  {% endblock %}
