+++ saleor/static/backend/js/return_purchase/create/containers/ItemList.js
@@ -26,7 +26,7 @@ class ItemList extends Component {
             {this.props.items.results.map(obj => {
               return (
                 <tr className={`${obj.quantity === 0 && 'hidden'}`} key={obj.id}>
-                    <td onClick={() => { this.addCart(obj); } }>
+                    <td title="add to cart" onClick={() => { this.addCart(obj); } }>
                       <span className={`cursor-pointer btn btn-primary btn-sm ${obj.quantity === 0 && 'hidden'}`}>
                         <i className="icon-cart-add"></i>
                       </span>
